Zad 1_1
